Move Molecule and enums to utils.models
Use Strandedness and Topology enums in multiple places
