+++ virtool_cli/ref/resources.py
@@ -3,7 +3,8 @@ from dataclasses import dataclass
 from pydantic import BaseModel
 from uuid import UUID
 
-from virtool_cli.ref.utils import DataType, IsolateName, Molecule
+from virtool_cli.ref.utils import DataType, IsolateName
+from virtool_cli.utils.models import Molecule
 
 
 class RepoMeta(BaseModel):
